connecting the sass generated css to the vue app
In the previous commit, I created the sass folder with the sass file and its output css file. Now I have connected the output file with the vue app (so that one will be actually used instead of the old one). However, the old one will be kept for reference. Also, a bug was already fixed that some background images were not showing properly.